renaming package name from movies to home
