+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
             android:theme="@style/AppTheme">
         <activity android:name=".ui.details.DetailsActivity">
         </activity>
-        <activity android:name=".ui.movies.MoviesActivity">
+        <activity android:name=".ui.home.MoviesActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
