+++ app/src/main/java/com/arturkida/popularmovies_kotlin/adapter/TabsAdapter.kt
@@ -3,8 +3,8 @@ package com.arturkida.popularmovies_kotlin.adapter
 import android.support.v4.app.Fragment
 import android.support.v4.app.FragmentManager
 import android.support.v4.app.FragmentPagerAdapter
-import com.arturkida.popularmovies_kotlin.ui.movies.FavoriteFragment
-import com.arturkida.popularmovies_kotlin.ui.movies.PopularFragment
+import com.arturkida.popularmovies_kotlin.ui.home.FavoriteFragment
+import com.arturkida.popularmovies_kotlin.ui.home.PopularFragment
 
 class TabsAdapter(fragmentManager: FragmentManager,
                   private val tabCount: Int) : FragmentPagerAdapter(fragmentManager) {
